employe deshboard/requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
opencv-python==4.9.0.80
face-recognition==1.3.0
Pillow==10.2.0
fpdf2==2.7.8
matplotlib==3.8.2
seaborn==0.13.2
openpyxl==3.1.2
scikit-learn==1.4.0
altair==5.2.0